+++ src/test/java/de/alpharogroup/db/resource/bundles/domain/LanguageLocaleTest.java
@@ -32,41 +32,40 @@ import org.testng.annotations.Test;
 import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 
 /**
- * The unit test class for the class {@link BaseName}
+ * The unit test class for the class {@link LanguageLocale}
  */
-public class BaseNameTest
+public class LanguageLocaleTest
 {
 
 	/**
-	 * Test method for {@link BaseName} constructors and builders
+	 * Test method for {@link LanguageLocale} constructors and builders
 	 */
 	@Test
 	public final void testConstructors()
 	{
-		BaseName model = new BaseName();
+		LanguageLocale model = new LanguageLocale();
 		assertNotNull(model);
-		model = BaseName.builder().build();
+		model = LanguageLocale.builder().build();
 		assertNotNull(model);
 	}
 
 	/**
-	 * Test method for {@link BaseName}
+	 * Test method for {@link LanguageLocale}
 	 */
 	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(BaseName.class);
+		beanTester.testBean(LanguageLocale.class);
 	}
 
 	/**
-	 * Test method for {@link BaseName#equals(Object)} , {@link BaseName#hashCode()} and
-	 * {@link BaseName#toString()}
+	 * Test method for {@link LanguageLocale#equals(Object)} , {@link LanguageLocale#hashCode()} and
+	 * {@link LanguageLocale#toString()}
 	 */
 	@Test
 	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		ContractVerifier.of(BaseName.class).verify();
+		ContractVerifier.of(LanguageLocale.class).verify();
 	}
-
 }