+++ src/test/java/de/alpharogroup/db/resource/bundles/domain/PropertiesValueTest.java
@@ -32,41 +32,40 @@ import org.testng.annotations.Test;
 import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 
 /**
- * The unit test class for the class {@link BaseName}
+ * The unit test class for the class {@link PropertiesValue}
  */
-public class BaseNameTest
+public class PropertiesValueTest
 {
 
 	/**
-	 * Test method for {@link BaseName} constructors and builders
+	 * Test method for {@link PropertiesValue} constructors and builders
 	 */
 	@Test
 	public final void testConstructors()
 	{
-		BaseName model = new BaseName();
+		PropertiesValue model = new PropertiesValue();
 		assertNotNull(model);
-		model = BaseName.builder().build();
+		model = PropertiesValue.builder().build();
 		assertNotNull(model);
 	}
 
 	/**
-	 * Test method for {@link BaseName}
+	 * Test method for {@link PropertiesValue}
 	 */
 	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(BaseName.class);
+		beanTester.testBean(PropertiesValue.class);
 	}
 
 	/**
-	 * Test method for {@link BaseName#equals(Object)} , {@link BaseName#hashCode()} and
-	 * {@link BaseName#toString()}
+	 * Test method for {@link PropertiesValue#equals(Object)} , {@link PropertiesValue#hashCode()}
+	 * and {@link PropertiesValue#toString()}
 	 */
 	@Test
 	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		ContractVerifier.of(BaseName.class).verify();
+		ContractVerifier.of(PropertiesValue.class).verify();
 	}
-
 }