+++ src/test/java/de/alpharogroup/db/resource/bundles/domain/ResourcebundleTest.java
@@ -32,41 +32,40 @@ import org.testng.annotations.Test;
 import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 
 /**
- * The unit test class for the class {@link BaseName}
+ * The unit test class for the class {@link Resourcebundle}
  */
-public class BaseNameTest
+public class ResourcebundleTest
 {
 
 	/**
-	 * Test method for {@link BaseName} constructors and builders
+	 * Test method for {@link Resourcebundle} constructors and builders
 	 */
 	@Test
 	public final void testConstructors()
 	{
-		BaseName model = new BaseName();
+		Resourcebundle model = new Resourcebundle();
 		assertNotNull(model);
-		model = BaseName.builder().build();
+		model = Resourcebundle.builder().build();
 		assertNotNull(model);
 	}
 
 	/**
-	 * Test method for {@link BaseName}
+	 * Test method for {@link Resourcebundle}
 	 */
 	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(BaseName.class);
+		beanTester.testBean(Resourcebundle.class);
 	}
 
 	/**
-	 * Test method for {@link BaseName#equals(Object)} , {@link BaseName#hashCode()} and
-	 * {@link BaseName#toString()}
+	 * Test method for {@link Resourcebundle#equals(Object)} , {@link Resourcebundle#hashCode()} and
+	 * {@link Resourcebundle#toString()}
 	 */
 	@Test
 	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		ContractVerifier.of(BaseName.class).verify();
+		ContractVerifier.of(Resourcebundle.class).verify();
 	}
-
 }